requirements.txt
transformers>=4.48.0
datasets>=3.0.0
torch>=2.8.0
accelerate>=1.0.0
peft>=0.15.0
trl>=0.12.0
openai>=1.108.1
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pandas>=2.2.0
numpy>=2.1.0